refactor(api): move GET top to asso visits service + rename it
